Move get_stored_modules to analysis_db
